feat: moved CountryFlag to where it should be
